fix variable scoping error
Due to the way the variable used in `except <exception> as <variable>`
is handled [1], it won't be in the scope of a function defined with
the `except` block.  It is necessary to first save the value to
another variable. Without this, execution that hits these except
blocks will fail with:

    NameError: name 'e' is not defined

[1] https://docs.python.org/3/reference/compound_stmts.html#the-try-statement